implement the new way of tracking motion
This is a lot less accurate than the previous way and it does not
track specific faces, only motion. when executing this the first
frame needs to be the background, so try to not be there when it
starts because it might not have the results you'd like.
Hopefully this less accurate solution can run on a raspberry pi